Legacy commit setup
The legacy version of this code was built in a rather janky fashion,
including the entirety of the bot being in one long file, the token
directly written into the file. It's also effectively whatever I could
think of thrown into one big file.

Just to keep it around, I'll be keeping this legacy code in the legacy
branch. I'll also be modifying the legacy code minimally so that the ID
and token are stored in their own files. It's a crude solution, but if
for some bizzare reason someone wants to get it working it should make
the work minimal.

Also, there's a lot of Dr. Strangelove references, because I decided to
use the character as the bot's profile picture when it was created.